+++ src/models/layout-page-element.ts
@@ -1,0 +1,9 @@
+// import type { LayoutPage } from '@app/models/layout-page';
+
+// class LayoutPageElement {
+//   declare layoutPage: LayoutPage;
+
+//   constructor() {
+//     // do something
+//   }
+// }
